backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy==2.0.36
python-docx==1.1.2
pdfplumber==0.11.4
google-genai
scikit-learn==1.5.2
numpy==2.1.3
python-dotenv==1.0.1
aiofiles==24.1.0
exa-py==1.0.2
